Changed mutation chance Implemented tournament selection Modified mutation code Updated fitness function and scoring system to use lines made as opposed to actual game score due to some scores rolling over into negative values

+++ TetrisMachineLearning/AIMain.h
@@ -10,14 +10,14 @@ using namespace sf;
 class AIMain
 {
 private:
-	int* _score;
+	unsigned long* _score;
 	int _movementArray[2]; //{Rotations, Translation}
 	bool _gameBoard[BOARD_WIDTH][BOARD_HEIGHT];
 	BoardState* _currentState;
 	float _rowsClearedMod, _closedHolesMod, _boardMaxHeightMod, _landingHeightMod, _surfaceRoughnessMod;
 	Brick _nextPiece;
 public:
-	AIMain(int* score); //Add setnext piece
+	AIMain(unsigned long* score); //Add setnext piece
 	~AIMain();
 	
 	BoardState* GetCurrentState() { return _currentState; }
